Fix: Memory leak by failing to dispose kernel in unit tests for Local array arguments, plus code quality improvements  (refs #79 #81)
